5.4.3

+++ Plugins/grpc_plugin/Source/Grpc/Grpc.Build.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Net.NetworkInformation;
-using System.Text;
+//using System.Net.NetworkInformation;
+//using System.Text;
 using System.Text.RegularExpressions;
 using UnrealBuildTool;
 
@@ -17,7 +17,7 @@ public class Grpc : ModuleRules
             this.root = root;
         }
 
-        private string root;
+        public string root;
 
         public string exe => Path.Combine(root, "vcpkg.exe");
         public string installed => Path.Combine(root, "installed");
@@ -62,7 +62,7 @@ public class Grpc : ModuleRules
     }
 
     //private string mPluginPath;
-    private string mCorePath;
+    //private string mCorePath;
 
     //private string mVcpkgPath;
     //private string mVcpkgExe;
@@ -130,14 +130,14 @@ public class Grpc : ModuleRules
 
     private HashSet<string> parseDepencies(VcpkgPaths vcpkgPaths, string package)
     {
-        //Regex reg = new Regex(@"^([^:\[]*?)(?:\[[^\]]*?\])??:(?:(?: ([^ ,]*?),)*? ([^ ]*?))??\r$", RegexOptions.Multiline);
         Regex reg = new Regex(@"^([^* :\[\]]*)(?:\[[^\]]*\])?:(.*)", RegexOptions.Multiline);
         HashSet<string> Packages = new HashSet<string>();
-
-        string unparsed = runProgram(vcpkgPaths.basePaths.exe, "depend-info " + package + ":" + vcpkgPaths.tripletName, true);
+        
+        string unparsed = runProgram(vcpkgPaths.basePaths.exe, "--overlay-ports=" + Path.GetFullPath(Path.Combine(PluginDirectory, "Source", "overlay")) + " --vcpkg-root " + vcpkgPaths.basePaths.root + " depend-info " + package + ":" + vcpkgPaths.tripletName, true);
         //Console.WriteLine(unparsed);
-        MatchCollection MatchCollection = reg.Matches(unparsed);
+        unparsed = unparsed.Replace(":" + vcpkgPaths.tripletName, "");
 
+        MatchCollection MatchCollection = reg.Matches(unparsed);
         foreach (Match match in MatchCollection)
         {
             Packages.Add(match.Groups[1].Value);
@@ -172,6 +172,13 @@ public class Grpc : ModuleRules
         string[] files = Directory.GetFiles(vcpkgPaths.info);
         List<string> InstalledFiles = new List<string>();
 
+        Console.WriteLine("Packages: [");
+        foreach (string Package in packages)
+        {
+            Console.WriteLine(Package);
+        }
+        Console.WriteLine("]");
+
         foreach (string Package in packages)
         {
             Regex regex = new Regex("^" + Package + @"_[^_]*?_" + vcpkgPaths.tripletName + @"\.list$");
@@ -204,13 +211,13 @@ public class Grpc : ModuleRules
             throw new BuildException("Linking directory failed: " + src + " -> " + target);
     }
     */
-    private static void checkInstalled(string path, Action install)
+    private static void checkInstalled(string path, Action install, bool precheck = true, bool postcheck = true)
     {
-        if (File.Exists(path)) return;
+        if (precheck && File.Exists(path)) return;
 
         install();
 
-        if (!File.Exists(path))
+        if (postcheck && !File.Exists(path))
             throw new BuildException("Failed at install of: " + Path.GetFileName(path));
     }
 
@@ -291,7 +298,7 @@ public class Grpc : ModuleRules
         foreach (FileInfo file in new DirectoryInfo(generatedDir).GetFiles())
         {
             
-            if (file.Extension != ".cc")
+            if (!file.FullName.EndsWith(".cc"))
             {
                 //Console.WriteLine("Gen header: \t" + file.Name);
                 //file.CopyTo(Path.Combine(outSource, file.Name), true);
@@ -332,7 +339,9 @@ public class Grpc : ModuleRules
         var linesToKeep = File.ReadLines(paths.tripletFile).Where(line => !line.Contains("VCPKG_BUILD_TYPE"));
 
         File.WriteAllLines(tempFile, linesToKeep);
-        File.AppendAllLines(tempFile, new string[] { "set(VCPKG_BUILD_TYPE release)" });
+        File.AppendAllLines(tempFile, new string[] { 
+            "set(VCPKG_BUILD_TYPE release)"
+        });
         
         File.Move(tempFile, paths.tripletFile, true);
     }
@@ -358,19 +367,22 @@ public class Grpc : ModuleRules
         VcpkgPaths TargetPaths = new VcpkgPaths(mTargetTriplet, basePaths);
         
         List<string> Packages = new List<string>();
-        Packages.AddRange(new string[] { "grpc", "draco", "asio-grpc" });
+        Packages.AddRange(new string[] { "grpc", /*"draco",*/ "asio-grpc" });
 
         checkInstalled(basePaths.exe, () =>
         {
             Console.WriteLine("Bootstrapping vcpkg");
-            runProgram(basePaths.bootstrap, "");
-        });
+            runProgram(basePaths.bootstrap, "-disableMetrics");
+        }, false);
+
+        string VcpkgCmd0 = "install --recurse --overlay-ports=" + Path.GetFullPath(Path.Combine(PluginDirectory, "Source", "overlay")) + " --host-triplet=" + mHostTriplet + " --vcpkg-root " + basePaths.root + " vcpkg-cmake";
+        runProgram(basePaths.exe, VcpkgCmd0);
 
         makeReleaseOnly(HostPaths);
         if (mHostTriplet != mTargetTriplet)
             makeReleaseOnly(TargetPaths);
 
-        string VcpkgCmd = "install --host-triplet=" + mHostTriplet;
+        string VcpkgCmd = "install --recurse --overlay-ports=" + Path.GetFullPath(Path.Combine(PluginDirectory, "Source", "overlay")) + " --host-triplet=" + mHostTriplet + " --vcpkg-root " + basePaths.root;
         string InstallMessage = "Installing [";
         HashSet<string> SubPackages = new HashSet<string>();
         foreach (var Package in Packages)
@@ -423,11 +435,12 @@ public class Grpc : ModuleRules
             if (i % 10 != 0)
                 Console.WriteLine("]");
         }
-        //Console.WriteLine();
+        
+        Console.WriteLine("Dlls:");
         foreach (string Dll in Ressources.dlls)
         {
             string DllFilePath = Path.Combine(TargetPaths.bin, Dll);
-            //Console.Write("\n" + Path.GetFileNameWithoutExtension(DllFilePath) + " ");
+            Console.Write("\n" + Path.GetFileNameWithoutExtension(DllFilePath) + " ");
             RuntimeDependencies.Add("$(TargetOutputDir)/" + Dll, DllFilePath);
         }
         //Console.WriteLine();
@@ -440,8 +453,7 @@ public class Grpc : ModuleRules
             "Core"
         });
 
-        if (Target.Platform == UnrealTargetPlatform.Win64)
-            PublicSystemLibraries.Add("crypt32.lib");
+        PublicSystemLibraries.Add("crypt32.lib");
 
         PrivateDependencyModuleNames.AddRange(new string[] { "CoreUObject", "Engine" });
     }
